moved GUI to separate lib
